Added a bulk mode option
When performing bulk inserts into alternative levelup backends (such as dynamodown or sqldown), the standard behavior that does a delete of old keys causes problems (massive reductions in performance when using sqldown, and total failure when using dynamodown because dynamo can't perform inserts and deletes within a single batch request). This modification provides an options object with (currently) a single memberthat can be passed in to disable the delete of the old key when performing a put.