Support the singular option when searching
Updated the geospatial search to support a new option ('singular') that is passed through to the underlying leveldown implementation. This is used with our SQLdown fork to ensure that the nested MAX(id) query isn't used for creating stream iterators (as it is incredibly slow when using massive database tables).